+++ src/main/java/com/elmakers/mine/bukkit/magic/MagicPlugin.java
@@ -163,6 +163,9 @@ public class MagicPlugin extends JavaPlugin implements MagicAPI
         TabExecutor magicGiveCommand = new MagicGiveCommandExecutor(this);
         getCommand("mgive").setExecutor(magicGiveCommand);
         getCommand("mgive").setTabCompleter(magicGiveCommand);
+        TabExecutor magicMapCommand = new MagicMapCommandExecutor(this);
+        getCommand("mmap").setExecutor(magicMapCommand);
+        getCommand("mmap").setTabCompleter(magicMapCommand);
         TabExecutor magicSkillsCommand = new MagicSkillsCommandExecutor(this);
         getCommand("mskills").setExecutor(magicSkillsCommand);
         getCommand("mskills").setTabCompleter(magicSkillsCommand);